Update Kmeans Functions.R
Added a function to pick AIC or BIC based on user entry.  This is partly working, I can get it to switch between the two depending on what I enter as "Criteria", but haven't gotten it to display the chosen one as a label.  Until I get that working I am still leaving in AIC and BIC (partly to make sure I was successfully switching between the two) and have not edited the plot portion.  I made comments with "NOTE" in them at the places where I need help or will eventually remove things.